+++ RecipeMatch/Views/DraggableView.m
@@ -156,7 +156,7 @@ static const float TITLE_X_OFFSET = 20;
             CGFloat rotationAngel = (CGFloat) (ROTATION_ANGLE * rotationStrength);
             
             // amount the height changes when you move the card up to a certain point
-            CGFloat scale = MAX(1 - fabsf(rotationStrength) / SCALE_STRENGTH, SCALE_MAX);
+            CGFloat scale = MAX(1 - fabs(rotationStrength) / SCALE_STRENGTH, SCALE_MAX);
             
             // move the object's center by center + gesture coordinate
             self.center = CGPointMake(self.originalPoint.x + xFromCenter, self.originalPoint.y + yFromCenter);
